+++ R/misc.R
@@ -21,13 +21,30 @@ get_rhs_vars <- function(formula, data, no_lhs = FALSE) {
   ## or should it? what about Y ~ log(x)?
   ## Answer: when called from `form2args`, the function
   ## `inline_check` stops when in-line functions are used.
-  data_info <- attr(model.frame(formula, data[1, ]), "terms")
-  response_info <- attr(data_info, "response")
-  predictor_names <- names(attr(data_info, "dataClasses"))
-  if (length(response_info) > 0 && all(response_info > 0)) {
-    predictor_names <- predictor_names[-response_info]
+
+  outcomes_names <- all.names(
+    rlang::f_lhs(formula), 
+    functions = FALSE, 
+    unique = TRUE
+  )
+
+  predictors_names <- all.names(
+    rlang::f_rhs(formula),
+    functions = FALSE,
+    unique = TRUE
+  )
+  
+  if (any(predictors_names == ".")) {
+    predictors_names <- predictors_names[predictors_names != "."]
+    predictors_names <- c(predictors_names, colnames(data))
+    predictors_names <- unique(predictors_names)
   }
-  predictor_names
+
+  if (length(predictors_names) > 0 && length(outcomes_names) > 0) {
+    predictors_names <- setdiff(predictors_names, outcomes_names)
+  }
+
+  predictors_names
 }
 
 #' Naming Tools
@@ -136,24 +153,9 @@ dummy_extract_names <- function(var, lvl, ordinal = FALSE, sep = "_") {
   nms
 }
 
-
-## As suggested by HW, brought in from the `pryr` package
-## https://github.com/hadley/pryr
-fun_calls <- function(f) {
-  if (is.function(f)) {
-    fun_calls(body(f))
-  } else if (is_quosure(f)) {
-    fun_calls(quo_get_expr(f))
-  } else if (is.call(f)) {
-    fname <- as.character(f[[1]])
-    # Calls inside .Internal are special and shouldn't be included
-    if (identical(fname, ".Internal")) {
-      return(fname)
-    }
-    unique(c(fname, unlist(lapply(f[-1], fun_calls), use.names = FALSE)))
-  }
+fun_calls <- function(f, data) {
+  setdiff(all.names(f), colnames(data))
 }
-
 
 get_levels <- function(x) {
   if (!is.factor(x) & !is.character(x)) {
@@ -191,9 +193,11 @@ strings2factors <- function(x, info) {
   for (i in seq_along(info)) {
     lcol <- names(info)[i]
     x[, lcol] <-
-      factor(as.character(x[[lcol]]),
+      factor(
+        as.character(x[[lcol]]),
         levels = info[[i]]$values,
-        ordered = info[[i]]$ordered
+        ordered = info[[i]]$ordered,
+        exclude = NULL
       )
   }
   x
@@ -262,11 +266,14 @@ merge_term_info <- function(.new, .old) {
 #' `ellipse_check()` is deprecated. Instead, empty selections should be
 #' supported by all steps.
 #'
-#' @param ... Arguments pass in from a call to `step`
-#' @return If not empty, a list of quosures. If empty, an error is thrown.
-#' @export
+#' @param ... Arguments pass in from a call to `step`.
+#' 
+#' @return `ellipse_check()`: If not empty, a list of quosures. If empty, an 
+#'   error is thrown.
+#' 
 #' @keywords internal
 #' @rdname recipes-internal
+#' @export
 ellipse_check <- function(...) {
   terms <- quos(...)
   if (is_empty(terms)) {
@@ -283,7 +290,7 @@ ellipse_check <- function(...) {
 
 #' Printing Workhorse Function
 #'
-#' This internal function is used for printing steps.
+#' `printer()` is used for printing steps.
 #'
 #' @param tr_obj A character vector of names that have been
 #'  resolved during preparing the recipe (e.g. the `columns` object
@@ -292,10 +299,12 @@ ellipse_check <- function(...) {
 #'  recipe (e.g. `terms` in most steps).
 #' @param trained A logical for whether the step has been trained.
 #' @param width An integer denoting where the output should be wrapped.
-#' @return `NULL`, invisibly.
+#' 
+#' @return `printer()`: `NULL`, invisibly.
+#' 
 #' @keywords internal
-#' @export
 #' @rdname recipes-internal
+#' @export
 printer <- function(tr_obj = NULL,
                     untr_obj = NULL,
                     trained = FALSE,
@@ -322,9 +331,9 @@ printer <- function(tr_obj = NULL,
 }
 
 
-#' @export
 #' @keywords internal
 #' @rdname recipes-internal
+#' @export
 prepare <- function(x, ...) {
   cli::cli_abort(
     "As of version 0.0.1.9006 please use {.fn prep} instead of {.fn prepare}."
@@ -481,12 +490,18 @@ check_type <- function(dat, quant = TRUE, types = NULL, call = caller_env()) {
 ## Support functions
 
 #' Check to see if a step or check as been trained
+#' 
+#' `is_trained()` is a helper function that returned a single logical to
+#' indicate whether a recipe is traine or not.
+#' 
 #' @param x a step object.
-#' @return A logical
-#' @export
-#' @keywords internal
-#' @rdname recipes-internal
+#' @return `is_trained()`: A single logical.
+#' 
 #' @seealso [developer_functions]
+#' @keywords internal
+#' 
+#' @rdname recipes-internal
+#' @export
 is_trained <- function(x) {
   x$trained
 }
@@ -494,15 +509,17 @@ is_trained <- function(x) {
 
 #' Convert Selectors to Character
 #'
-#' This internal function takes a list of selectors (e.g. `terms`
-#'  in most steps) and returns a character vector version for
-#'  printing.
+#' `sel2char()` takes a list of selectors (e.g. `terms` in most steps) and 
+#' returns a character vector version for printing.
+#' 
 #' @param x A list of selectors
-#' @return A character vector
-#' @export
+#' @return `sel2char()`: A character vector.
+#' 
+#' @seealso [developer_functions]
+#' 
 #' @keywords internal
 #' @rdname recipes-internal
-#' @seealso [developer_functions]
+#' @export
 sel2char <- function(x) {
   unname(map_chr(x, to_character))
 }
@@ -646,19 +663,12 @@ check_training_set <- function(x, rec, fresh, call = rlang::caller_env()) {
     }
     x <- rec$template
   } else {
-    in_data <- vars %in% colnames(x)
-    if (!all(in_data)) {
-      cli::cli_abort(
-        "Not all variables in the recipe are present in the supplied training \\
-        set: {.and {.var {vars[!in_data]}}}.",
-        call = call
-      )
-    }
     if (!is_tibble(x)) {
-      x <- as_tibble(x[, vars, drop = FALSE])
-    } else {
-      x <- x[, vars]
+      x <- as_tibble(x)
     }
+    recipes_ptype_validate(rec, new_data = x, stage = "prep", call = call)
+
+    x <- x[, vars]
   }
 
   steps_trained <- vapply(rec$steps, is_trained, logical(1))
@@ -937,4 +947,37 @@ recipes_remove_cols <- function(new_data, object, col_names = character()) {
     new_data <- new_data[, !(colnames(new_data) %in% removals), drop = FALSE]
   }
   new_data
+}
+
+#' Role indicators
+#'
+#' This helper function is meant to be used in `prep()` methods to identify
+#' predictors and outcomes by names.
+#'
+#' @param info data.frame with variable information of columns. 
+#'
+#' @return Character vector of column names.
+#' @keywords internal
+#'
+#' @seealso [developer_functions]
+#'
+#' @name recipes-role-indicator
+NULL
+
+#' @rdname recipes-role-indicator
+#' @export
+recipes_names_predictors <- function(info) {
+  get_from_info(info, "predictor")
+}
+
+#' @rdname recipes-role-indicator
+#' @export
+recipes_names_outcomes <- function(info) {
+  get_from_info(info, "outcome")
+}
+
+get_from_info <- function(info, role) {
+  res <- info$variable[info$role == role & !is.na(info$role)]
+
+  res
 }