+++ tests/testthat/test-dummy.R
@@ -196,8 +196,6 @@ test_that("tests for NA values in ordered factor", {
   )
 })
 
-
-
 test_that("new levels", {
   df <- data.frame(
     y = c(1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0),
@@ -215,10 +213,10 @@ test_that("new levels", {
   testing$x1 <- as.factor(testing$x1)
 
   expect_snapshot(
-    recipes:::warn_new_levels(testing$x1, levels(training$x1))
+    recipes:::warn_new_levels(testing$x1, levels(training$x1), "column", "step_dummy")
   )
   expect_silent(
-    recipes:::warn_new_levels(training$x1, levels(training$x1))
+    recipes:::warn_new_levels(training$x1, levels(training$x1), "column", "step_dummy")
   )
 
   rec <- recipe(y ~ x1, data = training) %>%
@@ -228,6 +226,16 @@ test_that("new levels", {
   )
   expect_snapshot(
     bake(rec, new_data = testing)
+  )
+})
+
+test_that("warns about NA in column (#450)", {
+  data <- data.frame(a = c(LETTERS, NA))
+
+  expect_snapshot(
+    tmp <- recipe(~ a, data = data) %>%
+      step_dummy(a) %>%
+      prep()
   )
 })
 
@@ -321,6 +329,18 @@ test_that("check_name() is used", {
   expect_snapshot(
     error = TRUE,
     prep(rec, training = dat)
+  )
+})
+
+test_that("throws a informative error for too many levels (#828)", {
+  dat <- data.frame(x = as.character(1:123456))
+
+  rec <- recipe(~ ., data = dat) %>%
+    step_dummy(x)
+
+  expect_snapshot(
+    error = TRUE,
+    prep(rec)
   )
 })